+++ realm/realm-library/src/main/cpp/io_realm_internal_TableQuery.cpp
@@ -831,7 +831,8 @@ enum StringPredicate {
     StringNotEqual,
     StringContains,
     StringBeginsWith,
-    StringEndsWith
+    StringEndsWith,
+    StringLike
 };
 
 
@@ -866,6 +867,9 @@ static void TableQuery_StringPredicate(JNIEnv *env, jlong nativeQueryPtr, jlongA
             case StringEndsWith:
                 Q(nativeQueryPtr)->ends_with(S(arr[0]), value2, is_case_sensitive);
                 break;
+            case StringLike:
+                Q(nativeQueryPtr)->like(S(arr[0]), value2, is_case_sensitive);
+                break;
             }
         }
         else {
@@ -885,6 +889,9 @@ static void TableQuery_StringPredicate(JNIEnv *env, jlong nativeQueryPtr, jlongA
                 break;
             case StringEndsWith:
                 Q(nativeQueryPtr)->and_query(table_ref->column<String>(size_t(arr[arr_len-1])).ends_with(StringData(value2), is_case_sensitive));
+                break;
+            case StringLike:
+                Q(nativeQueryPtr)->and_query(table_ref->column<String>(size_t(arr[arr_len-1])).like(StringData(value2), is_case_sensitive));
                 break;
             }
         }
@@ -913,6 +920,12 @@ JNIEXPORT void JNICALL Java_io_realm_internal_TableQuery_nativeEndsWith(
     JNIEnv* env, jobject, jlong nativeQueryPtr, jlongArray columnIndexes, jstring value, jboolean caseSensitive)
 {
     TableQuery_StringPredicate(env, nativeQueryPtr, columnIndexes, value, caseSensitive, StringEndsWith);
+}
+
+JNIEXPORT void JNICALL Java_io_realm_internal_TableQuery_nativeLike(
+        JNIEnv* env, jobject, jlong nativeQueryPtr, jlongArray columnIndexes, jstring value, jboolean caseSensitive)
+{
+    TableQuery_StringPredicate(env, nativeQueryPtr, columnIndexes, value, caseSensitive, StringLike);
 }
 
 JNIEXPORT void JNICALL Java_io_realm_internal_TableQuery_nativeContains(