+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/BaseIntegrationTest.java
@@ -20,21 +20,32 @@ import android.support.test.InstrumentationRegistry;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.mockito.internal.exceptions.ExceptionIncludingMockitoWarnings;
 
 import io.realm.Realm;
+import io.realm.log.RealmLog;
 import io.realm.objectserver.utils.HttpUtils;
 
-public abstract class RealmIntegrationTest {
+public class BaseIntegrationTest {
 
     @BeforeClass
     public static void setUp () throws Exception {
-        Realm.init(InstrumentationRegistry.getContext());
-        HttpUtils.startSyncServer();
+        try {
+            Realm.init(InstrumentationRegistry.getContext());
+            HttpUtils.startSyncServer();
+        } catch (Exception e) {
+            // Throwing an exception from this method will crash JUnit. Instead just log it.
+            // If this setup method fails, all unit tests in the class extending it will most likely fail as well.
+            RealmLog.error("Could not start Sync Server", e);
+        }
     }
 
     @AfterClass
     public static void tearDown () throws Exception {
-        HttpUtils.stopSyncServer();
+        try {
+            HttpUtils.stopSyncServer();
+        } catch (Exception e) {
+            RealmLog.error("Failed to stop Sync Server", e);
+        }
     }
-
 }