+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/utils/UserFactory.java
@@ -21,8 +21,19 @@ import io.realm.SyncUser;
 
 // Must be in `io.realm.objectserver` to work around package protected methods.
 public class UserFactory {
-    public static SyncUser createDefaultUser(String authUrl, String accessToken) {
-        SyncCredentials credentials = SyncCredentials.accessToken(accessToken, "sync-integration-user");
+
+    public static SyncUser createDefaultUser(String authUrl) {
+        return createUser(authUrl, "test-user");
+    }
+
+    public static SyncUser createUser(String authUrl, String userIdentifier) {
+        SyncCredentials credentials = SyncCredentials.usernamePassword(userIdentifier, "myPassw0rd", false);
+        return SyncUser.login(credentials, authUrl);
+    }
+
+    public static SyncUser createAdminUser(String authUrl) {
+        // `admin` required as user identifier to be granted admin rights.
+        SyncCredentials credentials = SyncCredentials.custom("admin", "debug", null);
         return SyncUser.login(credentials, authUrl);
     }
 }